requirements.txt
tavily-python
langchain
langchain_nvidia_ai_endpoints
langgraph
python-dotenv
openai